requirements.txt
torch>=2.0.0
torchvision>=0.15.2
numpy>=1.26.0
scipy>=1.11.1

Pillow>=9.4.0
imageio>=2.9.0
imageio-ffmpeg>=0.4.3
pyspng-seunglab>=1.1.2
gradio>=3.35.2
imgui==2.0.0
glfw==2.6.1
PyOpenGL>=3.1.5
matplotlib>=3.4.2

ninja==1.10.2

click>=8.1.0
requests>=2.26.0
tqdm>=4.62.2

# Optional: For model downloading (used by gradio/huggingface)
huggingface_hub>=1.2.3
hf_transfer>=0.1.9
